+++ gql/lib/src/language/parser.dart
@@ -609,9 +609,11 @@ class _Parser {
   }
 
   SchemaDefinitionNode _parseSchemaDefinition() {
+    final description = _parseDescription();
     _expectKeyword("schema");
 
     return SchemaDefinitionNode(
+      description: description,
       directives: _parseDirectives(
         isConst: true,
       ),
@@ -740,12 +742,14 @@ class _Parser {
     final description = _parseDescription();
     _expectKeyword("interface");
     final name = _parseName("Expected an interface name");
+    final interfaces = _parseImplementsInterfaces();
     final directives = _parseDirectives(isConst: true);
     final fields = _parseFieldsDefinition();
 
     return InterfaceTypeDefinitionNode(
       description: description,
       name: name,
+      interfaces: interfaces,
       directives: directives,
       fields: fields,
     );
@@ -881,6 +885,8 @@ class _Parser {
         return DirectiveLocation.fragmentSpread;
       case "INLINE_FRAGMENT":
         return DirectiveLocation.inlineFragment;
+      case "VARIABLE_DEFINITION":
+        return DirectiveLocation.variableDefinition;
       case "SCHEMA":
         return DirectiveLocation.schema;
       case "SCALAR":
@@ -988,6 +994,7 @@ class _Parser {
     _expectKeyword("interface");
 
     final name = _parseName("Expected an interface name");
+    final interfaces = _parseImplementsInterfaces();
     final directives = _parseDirectives(isConst: true);
     final fields = _parseFieldsDefinition();
 
@@ -1001,6 +1008,7 @@ class _Parser {
     return InterfaceTypeExtensionNode(
       name: name,
       directives: directives,
+      interfaces: interfaces,
       fields: fields,
     );
   }