+++ gql/lib/src/language/printer.dart
@@ -288,6 +288,10 @@ class _PrintVisitor extends Visitor<String> {
 
   @override
   String visitSchemaDefinitionNode(SchemaDefinitionNode node) => [
+        if (node.description != null) ...[
+          node.description!.accept(this),
+          "\n",
+        ],
         "schema",
         " ",
         visitDirectiveSetNode(node.directives),
@@ -435,6 +439,7 @@ class _PrintVisitor extends Visitor<String> {
         " ",
         node.name.accept(this),
         " ",
+        visitImplementsSetNode(node.interfaces),
         visitDirectiveSetNode(node.directives),
         visitFieldSetNode(node.fields),
       ].join();
@@ -661,6 +666,7 @@ class _PrintVisitor extends Visitor<String> {
         " ",
         node.name.accept(this),
         " ",
+        visitImplementsSetNode(node.interfaces),
         visitDirectiveSetNode(node.directives),
         visitFieldSetNode(node.fields),
       ].join();
@@ -736,6 +742,8 @@ String? _directiveLocation(DirectiveLocation location) {
       return "FRAGMENT_SPREAD";
     case DirectiveLocation.inlineFragment:
       return "INLINE_FRAGMENT";
+    case DirectiveLocation.variableDefinition:
+      return "VARIABLE_DEFINITION";
     case DirectiveLocation.schema:
       return "SCHEMA";
     case DirectiveLocation.scalar: