+++ links/gql_dio_link/test/gql_dio_link_test.dart
@@ -1,6 +1,5 @@
 // ignore_for_file: implicit_dynamic_method, unused_local_variable,implicit_dynamic_map_literal
 
-import "dart:async";
 import "dart:core";
 
 import "package:collection/collection.dart";