+++ gql/CHANGELOG.md
@@ -1,3 +1,8 @@
+## 0.13.1
+
+- Support `interfaces`.
+- Add support for Schema description.
+
 ## 0.13.0-nullsafety.2
 
 Loosen `source_span` constraint for `flutter_test` compatibility